+++ app/src/main/java/com/manna/fcm/MyFirebaseMessagingService.kt
@@ -4,10 +4,10 @@ import com.google.android.gms.tasks.OnCompleteListener
 import com.google.firebase.messaging.FirebaseMessaging
 import com.google.firebase.messaging.FirebaseMessagingService
 import com.google.firebase.messaging.RemoteMessage
-import com.manna.DeviceUtil
 import com.manna.MannaApp
 import com.manna.common.Logger
 import com.manna.di.ApiModule
+import com.manna.util.DeviceUtil
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.schedulers.Schedulers
