+++ app/src/main/java/com/manna/network/model/location/LocationResponse.kt
@@ -1,21 +1,19 @@
-package com.manna.presentation.chat
-
+package com.manna.network.model.location
 
 import com.google.gson.annotations.SerializedName
-import com.manna.Message
 import com.manna.network.model.chat.Sender
 
-data class ChatResponse(
+data class LocationResponse(
+    @SerializedName("location")
+    val latLng: MyLatLng?,
     @SerializedName("sender")
     val sender: Sender?,
-    @SerializedName("message")
-    val message: Message?,
     @SerializedName("type")
     val type: Type?
 ) {
     enum class Type {
-        @SerializedName("CHAT")
-        CHAT,
+        @SerializedName("LOCATION")
+        LOCATION,
         @SerializedName("JOIN")
         JOIN,
         @SerializedName("LEAVE")
