+++ app/src/main/java/com/manna/presentation/dialog/DateNotationDialog.kt
@@ -1,4 +1,4 @@
-package com.manna
+package com.manna.presentation.dialog
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,6 +6,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.DialogFragment
+import com.manna.R
 import com.manna.databinding.DialogDateNotationBinding
 
 class DateNotationDialog : DialogFragment() {
