+++ app/src/main/java/com/manna/presentation/sign_up/ServiceTermsFragment.kt
@@ -6,10 +6,10 @@ import android.text.method.ScrollingMovementMethod
 import android.view.View
 import androidx.activity.OnBackPressedCallback
 import androidx.core.content.ContextCompat
-import com.manna.HomeActivity
 import com.manna.R
 import com.manna.common.BaseFragment
 import com.manna.databinding.FragmentServiceTermsBinding
+import com.manna.presentation.HomeActivity
 import dagger.hilt.android.AndroidEntryPoint
