+++ app/src/main/res/layout/activity_setting.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".SettingActivity">
+        tools:context=".presentation.settings.SettingActivity">
 
 
         <androidx.constraintlayout.widget.ConstraintLayout
